requirements.txt
requests
python-dotenv
PyYAML